% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Eula.Seurat-package}
\alias{Eula.Seurat}
\alias{Eula.Seurat-package}
\title{Eula.Seurat: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
